+++ CHANGELOG.md
@@ -53,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add documentation around `DeviceSelection` and `useAudioInputActivityPreview` hook
 - Add docs for all icon components
 - Added documentation for Modal component again, after being reverted
+- Add RosterAttendee component, connected MicVolumeIndicator
 
 ### Changed
 
@@ -102,6 +103,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactor styles from SDK Component `DeviceSelection` and move them to demo
 - Moved `ActivityBar` from library to demo, since it seemed more tightly related our demo
 - Updated package.json to state sdk as peer dependency
+- Stop local video, if enabled, when selecting "none" as video input source
+- Add volume indicator, stop local video when selecting none as video source
 
 ### Removed